Sale and Customer handlers moved to separate directories
