+++ domain/customer/CustomerRepositoryStub.go
