+++ service/salesService.go
@@ -1,8 +1,9 @@
 package service
 
-import "github.com/oscaralmgren/rest-banking-api-go-tutorial/domain"
+import domain "github.com/oscaralmgren/rest-banking-api-go-tutorial/domain/sale"
 
-type SalesService interface {
+// Interface definition
+type SaleService interface {
 	FindOne() (*domain.Sale, error)
 }
 
@@ -10,14 +11,15 @@ type DefaultSaleService struct {
 	repo domain.SaleRepository
 }
 
+func NewSaleService(repositry domain.SaleRepository) DefaultSaleService {
+	return DefaultSaleService{repositry}
+}
+
+// Interface implementations from here.
 func (s DefaultSaleService) FindOne() (*domain.Sale, error) {
 	sale, err := s.repo.FindOne()
 	if err != nil {
 		return nil, err
 	}
 	return sale, nil
-}
-
-func NewSaleService(repositry domain.SaleRepository) DefaultSaleService {
-	return DefaultSaleService{repositry}
 }
